requirements.txt
# Flask Framework
Flask

# Flask Packages
Flask-Login
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
Flask-WTF
gunicorn